# internal/server/openapi.yaml
openapi: 3.1.0
info:
  title: Varkiv HTTP API
  summary: Stable, versioned contract for the Varkiv library and device ecosystem
  version: 0.1.0-preview.1
  description: |
    Versioned HTTP API for a self-hosted personal ROM catalog, frontend imports
    and exports, scoped device pairing, multi-file save streams, and automatic
    sync sessions. Every API response includes `X-Request-ID`,
    `X-Varkiv-API-Version`, and `Cache-Control: no-store`.

    Unless an operation explicitly declares `security: []`, deployments with an
    owner token require `Authorization: Bearer <token>`. Device tokens are opaque,
    independently revocable credentials whose effective scopes are enforced by
    the server. Signed content/session URLs are short-lived capabilities rather
    than bearer tokens.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
...
servers:
  - url: /api/v1
    description: Same-origin stable API entry point
security:
  - bearerAuth: []
...
paths:
...
  /games:
    get:
      operationId: getGames
      tags: [Library]
      summary: List games with editions and artifacts
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
        - {$ref: '#/components/parameters/Locale'}
        - {$ref: '#/components/parameters/Projection'}
        - {name: q, in: query, schema: {type: string}}
        - {name: platform, in: query, schema: {type: string}}
      responses:
        '200': {description: Paginated full or summary games, content: {application/json: {schema: {oneOf: [{$ref: '#/components/schemas/GameCollection'}, {$ref: '#/components/schemas/GameSummaryCollection'}]}}}}
        '400': {$ref: '#/components/responses/Error'}
    post:
      operationId: postGames
      tags: [Library]
      summary: Create a game
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/GameMutation'}}}}
      responses:
        '201': {description: Created game, content: {application/json: {schema: {$ref: '#/components/schemas/Game'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/ImportConflict'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Owner token or a server-issued device token; device scopes are enforced per route.
  headers:
    RequestID:
      description: Caller-supplied valid request identity or a server-generated identity.
      schema: {type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9._-]+$'}
    APIVersion:
      description: Stable API path version used for this response.
      schema: {type: string, const: v1}
  parameters:
...
    Locale: {name: locale, in: query, schema: {type: string, enum: [zh-CN, zh-TW, ja, en]}, description: Preferred localized title. The resource default remains the fallback.}
    Projection: {name: projection, in: query, schema: {type: string, enum: [full, summary], default: full}, description: "Use summary for list metadata, aggregate artifact state, and an opaque cover reference; fetch the detail resource before editing or launching."}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 100}, description: Maximum number of records returned in this page.}
    Offset: {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}, description: Zero-based record offset in the endpoint's documented stable ordering.}
...
  responses:
    ImportConflict:
      description: import_preview_stale when the signed snapshot no longer matches; import_batch_conflict when an atomic batch cannot commit; platform_definition_conflict/platform_definition_disabled/platform_key_conflict when a portable custom platform cannot be safely reused or created
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestID'}
        X-Varkiv-API-Version: {$ref: '#/components/headers/APIVersion'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Error:
      description: Structured API error
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestID'}
        X-Varkiv-API-Version: {$ref: '#/components/headers/APIVersion'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
          example:
            error:
              code: invalid_argument
              message: request payload is invalid
              request_id: 40e39da4b75c41c28cf40b9f8de46c48
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: {type: string, description: Stable machine-readable error code used for client branching.}
            message: {type: string, description: Human-readable diagnostic; clients must not parse this value.}
            request_id: {type: string, description: Correlation identity also returned in the X-Request-ID header.}
    Pagination:
      type: object
      additionalProperties: false
      description: Offset page metadata. Repeat the same filters and ordering while advancing offset by the number of returned records.
      required: [limit, offset, total]
      properties:
        limit: {type: integer, minimum: 1, maximum: 200}
        offset: {type: integer, minimum: 0}
        total: {type: integer, minimum: 0}
    GameMutation:
      type: object
      additionalProperties: false
      required: [default_title, platform]
      properties:
        id: {type: string, description: Optional caller-generated identity on create; ignored on update.}
        default_title: {type: string, minLength: 1}
        platform: {type: string, minLength: 1, description: Canonical platform ID or registered alias.}
        titles: {type: object, additionalProperties: {type: string}}
...
    Artifact:
      type: object
      required: [id, edition_id, path, storage_kind, role, size, missing]
      properties:
        id: {type: string, format: uuid}
        edition_id: {type: string, format: uuid}
        path: {type: string}
        storage_kind: {type: string, enum: [library, managed]}
        source_path: {type: string}
        original_name: {type: string}
        role: {type: string, enum: [rom, disc, executable, patch, dlc, update, other]}
        disc_index: {type: integer, minimum: 0, maximum: 64}
        size: {type: integer, format: int64, minimum: 0}
        sha256: {type: string}
        missing: {type: boolean}
...
    MediaAsset:
      type: object
      required: [id, kind, storage_kind, path, original_name, mime_type, size, sha256, source_type, content_status, created_at]
      properties:
        id: {type: string, format: uuid}
        game_id: {type: string, format: uuid}
        edition_id: {type: string, format: uuid}
        kind: {type: string}
        storage_kind: {type: string, enum: [library, managed]}
        path: {type: string}
        source_path: {type: string}
        original_name: {type: string}
        mime_type: {type: string}
        size: {type: integer, format: int64}
        sha256: {type: string}
        locale: {type: string}
        source_type: {type: string, enum: [upload, frontend-import, provider]}
        sort_order: {type: integer}
        content_status: {type: string, enum: [unverified, available, missing, changed, unsafe]}
        content_checked_at:
          description: RFC 3339 time of the last explicit check, or an empty string when the relation has not been checked yet.
          anyOf:
            - {type: string, format: date-time}
            - {type: string, const: ''}
        created_at: {type: string, format: date-time}
...
    Edition:
      type: object
      required: [id, game_id, default_title, edition_type, artifacts]
      properties:
        id: {type: string, format: uuid}
        game_id: {type: string, format: uuid}
        default_title: {type: string}
        display_title: {type: string}
        titles: {type: object, additionalProperties: {type: string}}
        edition_type: {type: string, enum: [original, translation, hack, revision, homebrew, other]}
        version: {type: string}
        author: {type: string}
        languages: {type: array, items: {type: string}}
        artifacts: {type: array, items: {$ref: '#/components/schemas/Artifact'}}
        media: {type: array, items: {$ref: '#/components/schemas/MediaAsset'}}
    Game:
      type: object
      required: [id, platform, default_title, editions]
      properties:
        id: {type: string, format: uuid}
        platform: {type: string}
        default_title: {type: string}
        display_title: {type: string}
        titles: {type: object, additionalProperties: {type: string}}
        primary_edition_id: {type: string, format: uuid}
        editions: {type: array, items: {$ref: '#/components/schemas/Edition'}}
        media: {type: array, items: {$ref: '#/components/schemas/MediaAsset'}}
    ArtifactStats:
      type: object
      additionalProperties: false
      required: [total, missing, hashed, usable, managed]
      properties:
        total: {type: integer, minimum: 0}
        missing: {type: integer, minimum: 0}
        hashed: {type: integer, minimum: 0}
        usable: {type: integer, minimum: 0, description: Non-missing artifacts with a content hash.}
        managed: {type: integer, minimum: 0}
    MediaSummary:
      type: object
      additionalProperties: false
      required: [id, kind, mime_type]
      description: Opaque thumbnail reference without storage paths, hashes, or source metadata.
      properties:
        id: {type: string, format: uuid}
        kind: {type: string}
        mime_type: {type: string}
    EditionSummary:
      type: object
      additionalProperties: false
      required: [id, game_id, default_title, display_title, edition_type, languages, save_namespace, sort_order, titles, artifact_stats]
      properties:
        id: {type: string, format: uuid}
        game_id: {type: string, format: uuid}
        default_title: {type: string}
        display_title: {type: string}
        edition_type: {type: string, enum: [original, translation, hack, revision, homebrew, other]}
        version: {type: string}
        languages: {type: array, items: {type: string}}
        author: {type: string}
        save_namespace: {type: string}
        serial: {type: string}
        product_code: {type: string}
        title_id: {type: string}
        sort_order: {type: integer}
        titles: {type: object, additionalProperties: {type: string}}
        artifact_stats: {$ref: '#/components/schemas/ArtifactStats'}
    GameSummary:
      type: object
      additionalProperties: false
      required: [id, default_title, display_title, platform, titles, editions, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        default_title: {type: string}
        display_title: {type: string}
        platform: {type: string}
        primary_edition_id: {type: string, format: uuid}
        titles: {type: object, additionalProperties: {type: string}}
        editions: {type: array, items: {$ref: '#/components/schemas/EditionSummary'}}
        cover: {$ref: '#/components/schemas/MediaSummary'}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    GameCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/Game'}}}}
    GameSummaryCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/GameSummary'}}}}
...
    PaginatedBase:
      type: object
      required: [data, pagination]
      properties:
        data: {type: array, items: {type: object}}
        pagination: {$ref: '#/components/schemas/Pagination'}